feat : cs read testcode 추가 및 레이어 변경되면서 테스트 구조 같이 변경
