--- src/test/java/com/workhub/cs/service/CsPostServiceTest.java
@@ -1,376 +1,0 @@
-package com.workhub.cs.service;
-
-import com.workhub.cs.dto.CsPostFileRequest;
-import com.workhub.cs.dto.CsPostRequest;
-import com.workhub.cs.dto.CsPostResponse;
-import com.workhub.cs.dto.CsPostUpdateRequest;
-import com.workhub.cs.entity.CsPost;
-import com.workhub.cs.entity.CsPostStatus;
-import com.workhub.cs.repository.CsPostFileRepository;
-import com.workhub.cs.repository.CsPostRepository;
-import com.workhub.global.error.ErrorCode;
-import com.workhub.global.error.exception.BusinessException;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.test.util.ReflectionTestUtils;
-
-import java.time.LocalDateTime;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.argThat;
-import static org.mockito.Mockito.*;
-
-@ExtendWith(MockitoExtension.class)
-public class CsPostServiceTest {
-
-    @Mock
-    private CsPostRepository csPostRepository;
-
-    @Mock
-    private CsPostFileRepository csPostFileRepository;
-
-    @InjectMocks
-    private CsPostService csPostService;
-
-    private CsPost mockSaved;
-
-    @BeforeEach
-    public void init(){
-        mockSaved = CsPost.builder()
-                .csPostId(1L)
-                .projectId(1L)
-                .userId(2L)
-                .title("문의 제목")
-                .content("문의 내용")
-                .build();
-    }
-
-    @Test
-    @DisplayName(" CS 게시글을 작성하면 작성한 게시글 정보를 보여준다.")
-    void givenCsPostCreateRequest_whenCreateCsPost_thenSuccess() {
-        // given
-        Long projectId = 1L;
-        Long userId = 2L;
-        CsPostRequest request = new CsPostRequest("문의 제목", "문의 내용", null);
-
-        when(csPostRepository.save(any(CsPost.class)))
-                .thenReturn(mockSaved);
-
-        // when
-        CsPostResponse result = csPostService.create(projectId, userId, request);
-
-        // then
-        assertThat(result.csPostId()).isEqualTo(1L);
-        assertThat(result.title()).isEqualTo("문의 제목");
-        assertThat(result.content()).isEqualTo("문의 내용");
-
-        verify(csPostRepository).save(any(CsPost.class));
-        verify(csPostFileRepository, never()).saveAll(anyList());
-    }
-
-    @Test
-    @DisplayName("파일이 포함된 게시글 작성 시 파일도 저장된다.")
-    void givenRequestWithFiles_whenCreate_thenFilesAreSaved() {
-        // given
-        Long projectId = 1L;
-        Long userId = 2L;
-
-        List<CsPostFileRequest> fileRequests = Arrays.asList(
-                new CsPostFileRequest("url1",  1),
-                new CsPostFileRequest("url2",  2)
-        );
-
-        CsPostRequest request =
-                new CsPostRequest("문의 제목", "내용", fileRequests);
-
-        when(csPostRepository.save(any(CsPost.class)))
-                .thenReturn(mockSaved);
-
-        // when
-        csPostService.create(projectId, userId, request);
-
-        // then
-        verify(csPostRepository).save(any(CsPost.class));
-        verify(csPostFileRepository, times(1)).saveAll(anyList()); // saveAll만 호출됨
-    }
-
-    // todo : 프로젝트 밸리데이터 붙이고 나서 테스트 해야 함
-    @Test
-    @DisplayName("프로젝트가 존재하지 않으면 ProjectNotFoundException이 발생한다.")
-    void givenInvalidProject_whenCreate_thenThrowProjectNotFound() {
-
-    }
-
-    @Test
-    @DisplayName("요청 DTO가 매핑되어 엔티티로 저장되는지 검증한다.")
-    void givenRequest_whenCreate_thenEntityMappedSuccessfully() {
-        // given
-        Long projectId = 1L;
-        Long userId = 2L;
-
-        CsPostRequest request = new CsPostRequest("문의 제목", "문의 내용", null);
-        when(csPostRepository.save(any(CsPost.class)))
-                .thenReturn(mockSaved);
-
-        // when
-        csPostService.create(projectId, userId, request);
-
-        // then
-        verify(csPostRepository).save(argThat(post ->
-                post.getProjectId().equals(projectId) &&
-                        post.getUserId().equals(userId) &&
-                        post.getTitle().equals("문의 제목") &&
-                        post.getContent().equals("문의 내용")
-        ));
-    }
-
-    @Test
-    @DisplayName("CS 게시글 수정 시 정상 저장되는지 검증한다.")
-    void givenCsPostUpdateRequest_whenUpdate_thenSuccess() {
-
-        // given
-        Long projectId = 1L;
-        Long csPostId = 2L;
-        Long userId = 2L;
-
-        CsPost original = CsPost.builder()
-                .csPostId(csPostId)
-                .projectId(projectId)
-                .userId(userId)
-                .title("원래 제목")
-                .content("원래 내용")
-                .build();
-
-        LocalDateTime originalUpdatedAt = LocalDateTime.now().minusMinutes(1);
-        ReflectionTestUtils.setField(original, "updatedAt", originalUpdatedAt);
-
-        CsPostUpdateRequest request =
-                new CsPostUpdateRequest("수정 제목", "수정 완료", List.of());
-
-        CsPost updated = CsPost.builder()
-                .csPostId(csPostId)
-                .projectId(projectId)
-                .userId(userId)
-                .title("수정 제목")
-                .content("수정 완료")
-                .build();
-
-        when(csPostRepository.findById(csPostId))
-                .thenReturn(Optional.of(original));
-
-        when(csPostRepository.save(any(CsPost.class)))
-                .thenReturn(updated);
-
-        ArgumentCaptor<CsPost> captor = ArgumentCaptor.forClass(CsPost.class);
-
-        // when
-        CsPostResponse result = csPostService.update(projectId, csPostId, userId, request);
-
-        // then
-        assertThat(result.title()).isEqualTo("수정 제목");
-        assertThat(result.content()).isEqualTo("수정 완료");
-
-        verify(csPostRepository).save(captor.capture());
-        CsPost savedEntity = captor.getValue();
-
-        assertThat(savedEntity.getTitle()).isEqualTo("수정 제목");
-        assertThat(savedEntity.getContent()).isEqualTo("수정 완료");
-
-        verify(csPostRepository).findById(csPostId);
-    }
-
-    @Test
-    @DisplayName("CS POST 게시글이 정상적으로 삭제된다.")
-    void givenDeleteCsPost_whenDelete_thenSuccess() {
-        // given
-        Long projectId = 1L;
-        Long csPostId = 1L;
-
-        when(csPostRepository.findById(csPostId)).thenReturn(Optional.of(mockSaved));
-
-        // when
-        csPostService.delete(projectId, csPostId);
-
-        // then
-        assertThat(mockSaved.getDeletedAt()).isNotNull();
-        verify(csPostRepository).findById(csPostId);
-
-    }
-
-    @Test
-    @DisplayName("CS POST 게시글 삭제시 예외 처리")
-    void givenNotExistsCsPost_whenDelete_thenThrowNotFound() {
-        // given
-        Long projectId = 1L;
-        Long csPostId = 1L;
-
-        when(csPostRepository.findById(csPostId)).thenReturn(Optional.empty());
-
-        // when & then
-        assertThatThrownBy(() -> csPostService.delete(projectId, csPostId))
-                .isInstanceOf(BusinessException.class);
-    }
-
-    @DisplayName("게시글 작성자가 아닌 사용자가 수정하려고 하면 FORBIDDEN_CS_POST_UPDATE 예외가 발생한다.")
-    void givenDifferentUser_whenUpdate_thenThrowForbidden() {
-        // given
-        Long projectId = 1L;
-        Long csPostId = 2L;
-        Long authorId = 2L;
-        Long requesterId = 3L; // 다른 사용자
-
-        CsPost original = CsPost.builder()
-                .csPostId(csPostId)
-                .projectId(projectId)
-                .userId(authorId)
-                .title("원래 제목")
-                .content("원래 내용")
-                .build();
-
-        CsPostUpdateRequest request =
-                new CsPostUpdateRequest("수정 제목", "수정 완료", List.of());
-
-        when(csPostRepository.findById(csPostId))
-                .thenReturn(Optional.of(original));
-
-        // when & then
-        assertThatThrownBy(() -> csPostService.update(projectId, csPostId, requesterId, request))
-                .isInstanceOf(BusinessException.class)
-                .hasFieldOrPropertyWithValue("errorCode", ErrorCode.FORBIDDEN_CS_POST_UPDATE);
-
-        verify(csPostRepository).findById(csPostId);
-        verify(csPostRepository, never()).save(any(CsPost.class));
-
-    }
-
-    @Test
-    @DisplayName("CS 게시글의 진행 상태를 변경한다.")
-    void givenStatus_whenStatusUpdate_thenSuccess() {
-        // given
-        CsPostStatus status = CsPostStatus.RECEIVED;
-
-        CsPost mockSaved = CsPost.builder()
-                .csPostId(1L)
-                .title("title")
-                .content("content")
-                .projectId(1L)
-                .userId(1L)
-                .csPostStatus(CsPostStatus.RECEIVED)
-                .build();
-
-        when(csPostRepository.findById(1L)).thenReturn(Optional.of(mockSaved));
-
-        // when
-        csPostService.changeStatus(1L, 1L, status);
-
-        // then
-        assertThat(mockSaved.getCsPostStatus()).isEqualTo(status);
-        verify(csPostRepository).findById(1L);
-        verify(csPostRepository, never()).save(any(CsPost.class));
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 CS 게시글 상태 변경 시 예외 발생")
-    void givenInvalidPostId_whenStatusUpdate_thenThrowException() {
-        // given
-        when(csPostRepository.findById(1L)).thenReturn(Optional.empty());
-
-        // when & then
-        assertThatThrownBy(() -> csPostService.changeStatus(1L, 1L, CsPostStatus.RECEIVED))
-                .isInstanceOf(BusinessException.class)
-                .hasMessage(ErrorCode.NOT_EXISTS_CS_POST.getMessage());
-
-        verify(csPostRepository).findById(1L);
-        verify(csPostRepository, never()).save(any(CsPost.class));
-    }
-
-    @Test
-    @DisplayName("상태를 정상 변경하면 변경된 상태를 반환한다.")
-    void givenStatus_whenChangeStatus_thenReturnUpdatedStatus() {
-        // given
-        CsPost mockSaved = CsPost.builder()
-                .csPostId(1L)
-                .title("title")
-                .content("content")
-                .projectId(1L)
-                .userId(1L)
-                .csPostStatus(CsPostStatus.RECEIVED)
-                .build();
-
-        when(csPostRepository.findById(1L)).thenReturn(Optional.of(mockSaved));
-
-        // when
-        CsPostStatus result = csPostService.changeStatus(1L, 1L, CsPostStatus.COMPLETED);
-
-        // then
-        assertThat(result).isEqualTo(CsPostStatus.COMPLETED);
-        assertThat(mockSaved.getCsPostStatus()).isEqualTo(CsPostStatus.COMPLETED);
-        verify(csPostRepository).findById(1L);
-        verify(csPostRepository, never()).save(any());
-    }
-
-    @Test
-    @DisplayName("CS POST 게시글 상세 조회한다.")
-    void givenCsPostId_whenGetCsPost_thenSuccess() {
-        // given
-        Long csPostId = 1L;
-        Long projectId = 1L;
-
-        when(csPostRepository.findById(csPostId)).thenReturn(Optional.of(mockSaved));
-
-        // when
-        CsPostResponse response = csPostService.findCsPost(projectId, csPostId);
-
-        // then
-        verify(csPostRepository).findById(csPostId);
-        assertThat(response.content()).isEqualTo("문의 내용");
-        assertThat(response.title()).isEqualTo("문의 제목");
-
-    }
-
-    @Test
-    @DisplayName("다른 프로젝트의 CS POST 조회 시 예외 발생")
-    void givenWrongProjectId_whenGetCsPost_thenThrow() {
-        // given
-        Long csPostId = 1L;
-        Long projectId = 999L; // 잘못된 projectId
-
-        CsPost mockSaved = CsPost.builder()
-                .csPostId(csPostId)
-                .title("문의 제목")
-                .content("문의 내용")
-                .projectId(1L)
-                .userId(1L)
-                .csPostStatus(CsPostStatus.RECEIVED)
-                .build();
-
-        when(csPostRepository.findById(csPostId)).thenReturn(Optional.of(mockSaved));
-
-        // when & then
-        assertThatThrownBy(() -> csPostService.findCsPost(projectId, csPostId))
-                .isInstanceOf(BusinessException.class)
-                .hasMessage(ErrorCode.NOT_MATCHED_PROJECT_CS_POST.getMessage());
-    }
-
-    @Test
-    @DisplayName("CS POST 게시글 목록을 조회한다.")
-    void t2() {
-        // given
-
-        // when
-
-        // then
-
-    }
-}
